In situation of garbage Exif value, would throw exception. Also, minor code cleanup.

+++ Source/Components/ImageGlass.Heart/Helpers.cs
@@ -121,7 +121,7 @@ namespace ImageGlass.Heart
         /// </summary>
         /// <param name="orientationFlag">Orientation Flag</param>
         /// <returns></returns>
-        public static double GetOrientationDegree(int orientationFlag)
+        public static int GetOrientationDegree(int orientationFlag)
         {
             if (orientationFlag == 1)
                 return 0;
